Move existing logic into command
